+++ h1st/model/model.md
@@ -32,7 +32,7 @@ PredictiveModel -up-> Model
 MLModel -up->PredictiveModel
 BooleanModel -up->PredictiveModel
 RuleBasedModel -up->PredictiveModel
-OracleModel -up-> PredictiveModel
+Oracle -up-> PredictiveModel
 
 FuzzyModel -up->RuleBasedModel
 @enduml
