+++ h1st/model/oracle/__init__.py
@@ -1,4 +1,4 @@
 from h1st.model.oracle.oracle_modeler import OracleModeler
-from h1st.model.oracle.oracle_model import OracleModel
+from h1st.model.oracle.oracle import Oracle
 from h1st.model.oracle.ts_oracle_modeler import TimeseriesOracleModeler
 from h1st.model.oracle.ts_oracle_model import TimeSeriesOracleModel
